refactor(test): split goUsageTracker tests into focused suites with shared helpers
